Split Fabric-x client and endorser wiring
