+++ platform/fabricx/sdk/dig/client/providers.go
@@ -4,23 +4,28 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package sdk
+package client
 
 import (
+	"context"
+	"errors"
+
 	cb "github.com/hyperledger/fabric-protos-go-apiv2/common"
 	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/dig"
 
 	"github.com/hyperledger-labs/fabric-smart-client/platform/common/driver"
-	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core"
+	commondig "github.com/hyperledger-labs/fabric-smart-client/platform/common/sdk/dig"
+	digutils "github.com/hyperledger-labs/fabric-smart-client/platform/common/utils/dig"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic/delivery"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic/driver/config"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic/driver/identity"
 	fdriver "github.com/hyperledger-labs/fabric-smart-client/platform/fabric/driver"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/services/db/driver/multiplexed"
-	"github.com/hyperledger-labs/fabric-smart-client/platform/fabricx"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabricx/core/channel"
+	committerconfig "github.com/hyperledger-labs/fabric-smart-client/platform/fabricx/core/committer/config"
+	committergrpc "github.com/hyperledger-labs/fabric-smart-client/platform/fabricx/core/committer/grpc"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabricx/core/committer/queryservice"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabricx/core/finality"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabricx/core/ledger"
@@ -32,43 +37,45 @@ import (
 	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/storage/kvs"
 )
 
-type P2PCommunicationType = string
-
-const (
-	WebSocket P2PCommunicationType = "websocket"
-)
-
-func NewDriver(in struct {
-	dig.In
-	ConfigProvider  config.Provider
-	MetricsProvider metrics.Provider
-	EndpointService identity.EndpointService
-	IdProvider      identity.ViewIdentityProvider
-	KVS             *kvs.KVS
-	SignerInfoStore driver.SignerInfoStore
-	AuditInfoStore  driver.AuditInfoStore
-	ChannelProvider ChannelProvider
-	IdentityLoaders []identity.NamedIdentityLoader `group:"identity-loaders"`
-},
-) core.NamedDriver {
-	d := core.NamedDriver{
-		Name: fabricx.DriverName,
-		Driver: fabricx.NewProvider(
-			in.ConfigProvider,
-			in.MetricsProvider,
-			in.EndpointService,
-			in.ChannelProvider,
-			in.IdProvider,
-			in.IdentityLoaders,
-			in.KVS,
-			in.SignerInfoStore,
-			in.AuditInfoStore,
-		),
-	}
-	return d
+func Install(sdk commondig.SDK) error {
+	return errors.Join(
+		sdk.Container().Provide(committerconfig.NewProvider, dig.As(
+			new(committergrpc.ServiceConfigProvider),
+			new(finality.ServiceConfigProvider),
+			new(queryservice.ServiceConfigProvider),
+		)),
+		sdk.Container().Provide(committergrpc.NewClientProvider, dig.As(
+			new(ledger.GRPCClientProvider),
+			new(queryservice.GRPCClientProvider),
+			new(finality.GRPCClientProvider),
+		)),
+		sdk.Container().Provide(ledger.NewProvider),
+		sdk.Container().Provide(finality.NewListenerManagerProvider),
+		sdk.Container().Provide(digutils.Identity[*finality.Provider](), dig.As(new(finality.ListenerManagerProvider))),
+		sdk.Container().Provide(queryservice.NewProvider, dig.As(new(queryservice.Provider))),
+		sdk.Container().Provide(NewChannelProvider, dig.As(new(generic.ChannelProvider))),
+	)
 }
 
-type ChannelProvider generic.ChannelProvider
+func RegisterLegacy(sdk commondig.SDK) error {
+	return errors.Join(
+		digutils.Register[finality.ListenerManagerProvider](sdk.Container()),
+		digutils.Register[queryservice.Provider](sdk.Container()),
+		digutils.Register[*ledger.Provider](sdk.Container()),
+	)
+}
+
+func InitializeProviders(sdk commondig.SDK, ctx context.Context) error {
+	return sdk.Container().Invoke(func(in struct {
+		dig.In
+		FinalityProvider *finality.Provider
+		LedgerProvider   *ledger.Provider
+	}) error {
+		in.FinalityProvider.Initialize(ctx)
+		in.LedgerProvider.Initialize(ctx)
+		return nil
+	})
+}
 
 func NewChannelProvider(in struct {
 	dig.In
@@ -87,8 +94,7 @@ func NewChannelProvider(in struct {
 	MetadataStore           fdriver.MetadataStore
 	EndorseTxStore          fdriver.EndorseTxStore
 	Drivers                 multiplexed.Driver
-},
-) generic.ChannelProvider {
+}) generic.ChannelProvider {
 	channelConfigProvider := generic.NewChannelConfigProvider(in.ConfigProvider)
 	return channel.NewProvider(
 		in.ConfigProvider,
@@ -106,7 +112,6 @@ func NewChannelProvider(in struct {
 		func(channel string, nw fdriver.FabricNetworkService, envelopeService fdriver.EnvelopeService, transactionService fdriver.EndorserTransactionService, vault fdriver.RWSetInspector) (fdriver.RWSetLoader, error) {
 			return NewRWSetLoader(channel, nw, envelopeService, transactionService, vault), nil
 		},
-		// delivery service constructor
 		func(
 			nw fdriver.FabricNetworkService,
 			channel string,
@@ -135,7 +140,6 @@ func NewChannelProvider(in struct {
 				[]cb.HeaderType{cb.HeaderType_MESSAGE},
 			)
 		},
-		// membership service
 		func(channelName string) fdriver.MembershipService {
 			return membership.NewService(channelName)
 		},
